Update to OpenFMB 2.0
* Many f32 to f64 changes
* Cleanup warnings about unused imports
* Drop dead subscription threads from nats
* Updated examples with working switch state toggling

+++ openfmb-messages-ext/src/ess/mod.rs
@@ -9,7 +9,7 @@ pub use reading::*;
 pub use status::*;
 
 pub trait ESSStatusExt {
-    fn soc(&self) -> Result<f32, OpenFMBError>;
+    fn soc(&self) -> Result<f64, OpenFMBError>;
 }
 
 pub trait ESSReadingExt {}
